requirements.txt
requests~=2.32.5
pandas~=3.0.0
tqdm~=4.67.3
python-dotenv